Refactor project structure and update dependencies. Creating a pyproject.toml file.
