+++ vowspace/assets/vowspace-icon-raw.png
