+++ vowspace/core/normalization.py
