+++ vowspace/vowel_space_visualizer.py
@@ -1,27 +1,24 @@
 # vowel_space_visualizer.py
 # This is where everything else comes together.
 
+import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-import matplotlib.pyplot as plt
-import openpyxl
+from PyQt5.QtCore import Qt, QTimer
 from PyQt5.QtGui import QIcon
-from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
-from matplotlib.patches import Ellipse
-from scipy.spatial import ConvexHull
-from scipy.stats import chi2
 from PyQt5.QtWidgets import (
     QWidget, QLabel, QLineEdit, QPushButton, QVBoxLayout, QHBoxLayout,
     QGridLayout, QFileDialog, QMessageBox, QMenu, QMenuBar, QAction, QCheckBox, QComboBox
 )
-from PyQt5.QtCore import Qt, QTimer
-from PyQt5.QtGui import QIcon
+from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
+from matplotlib.patches import Ellipse
+from scipy.spatial import ConvexHull
+from scipy.stats import chi2
 
-from components.df_editor import DFEditor
-from components.ipa_window import IPAWindow
-from components.audio_tool import AudioAnalysisTool
-
-from core.normalization import (
+from vowspace.components.audio_tool import AudioAnalysisTool
+from vowspace.components.df_editor import DFEditor
+from vowspace.components.ipa_window import IPAWindow
+from vowspace.core.normalization import (
     lobanov_normalization,
     bark_difference,
     nearey1,
@@ -51,7 +48,7 @@ class VowelSpaceVisualizer(QWidget):
         # Set initial state
         self.data = pd.DataFrame(columns=["vowel", "f0", "f1", "f2", "f3", "f4", "speaker"])
         self.setWindowTitle("VowSpace v1.4.2")
-        self.setWindowIcon(QIcon("assets/vowspace.ico"))
+        self.setWindowIcon(QIcon("vowspace/assets/vowspace.ico"))
 
         self.create_menu_bar()
